SdkComponentFactoryTest: Create one class for each role
Instead of having the test class be also the factory and the component
implementation, create a separate class for each. This makes the test
easier to understand.
