+++ src/test/java/eu/europa/ted/eforms/sdk/component/MyComponent.java
@@ -1,0 +1,10 @@
+package eu.europa.ted.eforms.sdk.component;
+
+@SdkComponent(versions = "0.5", componentType = SdkComponentType.EFX_EXPRESSION_TRANSLATOR)
+class MyComponent implements TestComponent {
+
+  @Override
+  public String testMethod() {
+    return null;
+  }
+}
