+++ src/test/java/eu/europa/ted/eforms/sdk/component/MyComponentFactory.java
@@ -1,0 +1,10 @@
+package eu.europa.ted.eforms.sdk.component;
+
+class MyComponentFactory extends SdkComponentFactory {
+
+  @Override
+  protected <T> T getComponentImpl(String sdkVersion, SdkComponentType componentType, Class<T> intf,
+      Object... initArgs) throws InstantiationException {
+    return super.getComponentImpl(sdkVersion, componentType, intf, initArgs);
+  }
+}
